Auth is smooth, just need to do it automatically
