+++ .idea/misc.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="Black">
+    <option name="enabledOnSave" value="true" />
     <option name="sdkName" value="Poetry (ripper)" />
   </component>
   <component name="ProjectRootManager" version="2" project-jdk-name="Poetry (ripper)" project-jdk-type="Python SDK" />
