+++ ripperlib/database.py
