--- src/widgets/mainview.py
@@ -1,302 +1,0 @@
-import logging
-import sys
-
-from PySide6.QtCore import QFile, Qt, QTextStream, QSize
-from PySide6.QtGui import QAction, QFont, QIcon, QKeySequence
-from PySide6.QtPrintSupport import QPrintDialog, QPrinter
-from PySide6.QtWidgets import (QApplication, QDialog, QDockWidget,
-                               QFileDialog, QMainWindow,
-                               QMessageBox, QWidget, QGridLayout, QLabel)
-from src.auth import auth_manager, AuthState
-
-import res.images.tools
-from src.widgets.auth_view import AuthView
-
-
-class MainView(QMainWindow):
-    def __init__(self):
-        super().__init__()
-
-        self.log = logging.getLogger('ripper:mainview')
-
-        self._file_menu = None
-        self._edit_menu = None
-        self._view_menu = None
-        self._oauth_menu = None
-        self._help_menu = None
-
-        self._file_tool_bar = None
-        self._edit_tool_bar = None
-
-        self._register_oauth_act = None
-        self._authenticate_oauth_act = None
-        self._new_source_act = None
-        self._save_act = None
-        self._print_act = None
-        self._undo_act = None
-        self._quit_act = None
-        self._about_act = None
-        self._about_qt_act = None
-
-        self.setWindowTitle("ripper")
-        self.dockNestingEnabled = True
-        self.resize(QSize(640, 480))
-
-        self.create_actions()
-        self.create_menus()
-        self.create_tool_bars()
-        self.create_status_bar()
-
-        # Ensure we have a central widget, then hide it
-        grid_layout = QGridLayout()
-        widget = QWidget()
-        widget.setLayout(grid_layout)
-        self.setCentralWidget(widget)
-        self.centralWidget().hide()
-
-        # Check if OAuth client is configured and update UI accordingly
-        self.update_oauth_ui()
-
-    def create_menus(self):
-        self._file_menu = self.menuBar().addMenu("&File")
-        self._file_menu.addAction(self._new_source_act)
-        self._file_menu.addAction(self._save_act)
-        self._file_menu.addAction(self._print_act)
-        self._file_menu.addSeparator()
-        self._file_menu.addAction(self._quit_act)
-
-        self._edit_menu = self.menuBar().addMenu("&Edit")
-        self._edit_menu.addAction(self._undo_act)
-
-        self._view_menu = self.menuBar().addMenu("&View")
-
-        self._oauth_menu = self.menuBar().addMenu("&OAuth")
-        self._oauth_menu.addAction(self._register_oauth_act)
-        self._oauth_menu.addAction(self._authenticate_oauth_act)
-
-        self.menuBar().addSeparator()
-
-        self._help_menu = self.menuBar().addMenu("&Help")
-        self._help_menu.addAction(self._about_act)
-        self._help_menu.addAction(self._about_qt_act)
-
-    def create_tool_bars(self):
-        self._file_tool_bar = self.addToolBar("File")
-        self._file_tool_bar.addAction(self._new_source_act)
-        self._file_tool_bar.addAction(self._save_act)
-        self._file_tool_bar.addAction(self._print_act)
-
-        self._edit_tool_bar = self.addToolBar("Edit")
-        self._edit_tool_bar.addAction(self._undo_act)
-
-    def create_status_bar(self):
-        self.statusBar().showMessage("Ready")
-
-        # Create a permanent widget for auth status
-        self.auth_status_label = QLabel()
-        self.auth_status_label.setFrameStyle(QLabel.Panel | QLabel.Sunken)
-        self.auth_status_label.setMinimumWidth(200)
-        self.statusBar().addPermanentWidget(self.auth_status_label)
-
-        # Connect to auth state changed signal
-        auth_manager.authStateChanged.connect(self.update_auth_status)
-
-        # Initialize auth status display
-        self.update_auth_status(*auth_manager.get_current_state())
-
-
-    ####### Actions #####################################################################
-
-    # noinspection PyArgumentList
-    def create_actions(self):
-        # TODO add google oauth-specific icon
-        icon = QIcon.fromTheme('document-new', QIcon(':/res/new.png'))
-        self._register_oauth_act = QAction(icon, "Register/Update OAuth Client",
-                                           self, statusTip="Register or update the target Google OAuth Client",
-                                           triggered=self.register_oauth)
-
-        # Authenticate action
-        icon = QIcon.fromTheme('dialog-password', QIcon(':/res/new.png'))
-        self._authenticate_oauth_act = QAction(icon, "Authenticate",
-                                              self, statusTip="Start Google OAuth authentication flow",
-                                              triggered=self.authenticate_oauth)
-        # Initially disabled until client is configured
-        self._authenticate_oauth_act.setEnabled(False)
-
-        icon = QIcon.fromTheme('document-new', QIcon(':/res/new.png'))
-        self._new_source_act = QAction(icon, "&New Source",
-                                       self, shortcut=QKeySequence.New,
-                                       statusTip="Import a new source sheet",
-                                       triggered=self.new_source)
-
-
-        icon = QIcon.fromTheme('document-save', QIcon(':/res/save.png'))
-        self._save_act = QAction(icon, "&Save...", self,
-                                 shortcut=QKeySequence.Save,
-                                 statusTip="Save the current spreadsheet", triggered=self.save)
-
-        icon = QIcon.fromTheme('document-print', QIcon(':/res/print.png'))
-        self._print_act = QAction(icon, "&Print...", self,
-                                  shortcut=QKeySequence.Print,
-                                  statusTip="Print the current spreadsheet",
-                                  triggered=self.print_)
-
-        icon = QIcon.fromTheme('edit-undo', QIcon(':/res/undo.png'))
-        self._undo_act = QAction(icon, "&Undo", self,
-                                 shortcut=QKeySequence.Undo,
-                                 statusTip="Undo the last editing action", triggered=self.undo)
-
-        self._quit_act = QAction("&Quit", self, shortcut="Ctrl+Q",
-                                 statusTip="Quit the application", triggered=self.close)
-
-        self._about_act = QAction("&About", self,
-                                  statusTip="About ripper",
-                                  triggered=self.about)
-
-        self._about_qt_act = QAction("About &Qt", self,
-                                     statusTip="About Qt",
-                                     triggered=QApplication.instance().aboutQt)
-
-    # User is prompted to supply an OAuth client and authenticate with Google API
-    def register_oauth(self):
-        self.log.debug("Register OAuth selected")
-        self.show_auth_view()
-
-    # Start the Google OAuth authentication flow
-    def authenticate_oauth(self):
-        self.log.debug("Authenticate OAuth selected")
-
-        try:
-            # Show a message to inform the user about the browser opening
-            QMessageBox.information(self, "Google Authentication",
-                                   "A browser window will open for you to authenticate with Google.\n\n"
-                                   "Please follow the instructions in the browser.")
-
-            # Start the OAuth flow
-            cred = auth_manager.authorize()
-
-            if cred:
-                QMessageBox.information(self, "Authentication Successful",
-                                       "Successfully authenticated with Google!")
-            else:
-                QMessageBox.warning(self, "Authentication Failed",
-                                   "Failed to authenticate with Google. Please try again.")
-
-            # Auth state is already updated in the authorize() function
-        except Exception as e:
-            self.log.error(f"Error during OAuth authentication: {str(e)}")
-            QMessageBox.critical(self, "Authentication Error",
-                               f"An error occurred during authentication:\n{str(e)}")
-            # Update auth state in case of exception
-            auth_manager.update_auth_state()
-
-    # Check if OAuth client is configured and update UI accordingly
-    def update_oauth_ui(self):
-        """Check if OAuth client credentials are available and update UI accordingly"""
-        # Get current auth state
-        state, _ = auth_manager.get_current_state()
-        has_credentials = state != AuthState.NO_CLIENT
-
-        # Enable/disable the authenticate action based on whether credentials are available
-        if hasattr(self, '_authenticate_oauth_act') and self._authenticate_oauth_act:
-            self._authenticate_oauth_act.setEnabled(has_credentials)
-
-        self.log.debug(f"OAuth client credentials {'found' if has_credentials else 'not found'}, "
-                      f"authenticate action {'enabled' if has_credentials else 'disabled'}")
-
-    def new_source(self):
-        self.log.debug("New source selected")
-
-    def show_auth_view(self):
-        """Show the authentication view as a dialog"""
-        self.auth_dialog = QDialog(self)
-        self.auth_dialog.setWindowTitle("Google API Authentication")
-        self.auth_dialog.setMinimumWidth(500)
-
-        # Create auth view
-        auth_view = AuthView(self.auth_dialog)
-        auth_view.oauth_client_registered.connect(self.on_oauth_client_registered)
-
-        # Set layout
-        layout = QGridLayout(self.auth_dialog)
-        layout.addWidget(auth_view)
-        self.auth_dialog.setLayout(layout)
-
-        # Show dialog
-        self.auth_dialog.exec()
-
-    def print_(self):
-        document = self._text_edit.document()
-        printer = QPrinter()
-
-        dlg = QPrintDialog(printer, self)
-        if dlg.exec() != QDialog.Accepted:
-            return
-
-        document.print_(printer)
-
-        self.statusBar().showMessage("Ready", 2000)
-
-    def save(self):
-        dialog = QFileDialog(self, "Choose a file name")
-        dialog.setMimeTypeFilters(['text/html'])
-        dialog.setAcceptMode(QFileDialog.AcceptSave)
-        dialog.setDefaultSuffix('csv')
-        if dialog.exec() != QDialog.Accepted:
-            return
-
-        filename = dialog.selectedFiles()[0]
-        file = QFile(filename)
-        if not file.open(QFile.WriteOnly | QFile.Text):
-            reason = file.errorString()
-            QMessageBox.warning(self, "Dock Widgets",
-                                f"Cannot write file {filename}:\n{reason}.")
-            return
-
-        out = QTextStream(file)
-        with QApplication.setOverrideCursor(Qt.WaitCursor):
-            out << self._text_edit.toPlainText()
-
-        self.statusBar().showMessage(f"Saved '{filename}'", 2000)
-
-    def undo(self):
-        self.log.debug("Undo selected")
-
-
-    def about(self):
-        QMessageBox.about(self, "About ripper",
-                          "This is ripper")
-
-
-    ####### Slots #######################################################################
-
-    # Slot to update the auth status in the status bar
-    def update_auth_status(self, state, user_email=""):
-        """Update the auth status display in the status bar"""
-        if state == AuthState.NO_CLIENT:
-            self.auth_status_label.setText("No OAuth Client")
-        elif state == AuthState.NOT_LOGGED_IN:
-            self.auth_status_label.setText("Not Logged In")
-        elif state == AuthState.LOGGED_IN:
-            self.auth_status_label.setText(f"Logged In: {user_email}")
-        else:
-            self.auth_status_label.setText("Unknown Auth State")
-
-        # Update UI elements that depend on auth state
-        self.update_oauth_ui()
-
-    # Slot to be called when the user successfully selects or updates the target OAuth client
-    def on_oauth_client_registered(self):
-        """Handle user update of target oauth client"""
-        self.log.debug("User configured OAuth client credentials")
-
-        # Close the auth dialog
-        if hasattr(self, 'auth_dialog') and self.auth_dialog:
-            self.auth_dialog.accept()
-
-        # Update UI to enable options only available after an OAuth client is configured
-        self.update_oauth_ui()
-
-        # Show a message that authentication was successful
-        QMessageBox.information(self, "OAuth Client Update Successful",
-                               "OAuth Client credentials successfully updated and stored securely")
